+++ tests/Question/ChoiceQuestionTest.php
@@ -30,7 +30,7 @@ class ChoiceQuestionTest extends TestCase
             '/',
         );
 
-        [$response, $env] = $question($env);
+        [$response, $env] = $question($env)->unwrap();
         $response = $response->match(
             static fn($response) => $response,
             static fn() => null,
@@ -71,7 +71,7 @@ class ChoiceQuestionTest extends TestCase
             '/',
         );
 
-        [$response, $env] = $question($env);
+        [$response, $env] = $question($env)->unwrap();
 
         $this->assertNull($response->match(
             static fn($response) => $response,
@@ -91,7 +91,7 @@ class ChoiceQuestionTest extends TestCase
             '/',
         );
 
-        [$response, $env] = $question($env);
+        [$response, $env] = $question($env)->unwrap();
 
         $this->assertNull($response->match(
             static fn($response) => $response,